Move the kubelet's external types to k8s.io/kubelet
