+++ hack/.golint_failures
@@ -704,6 +704,7 @@ staging/src/k8s.io/kube-aggregator/pkg/controllers/status
 staging/src/k8s.io/kube-aggregator/pkg/registry/apiservice
 staging/src/k8s.io/kube-aggregator/pkg/registry/apiservice/etcd
 staging/src/k8s.io/kube-proxy/config/v1alpha1
+staging/src/k8s.io/kubelet/config/v1beta1
 staging/src/k8s.io/metrics/pkg/apis/custom_metrics
 staging/src/k8s.io/metrics/pkg/apis/custom_metrics/v1beta1
 staging/src/k8s.io/metrics/pkg/apis/custom_metrics/v1beta2
