+++ pkg/kubelet/apis/config/v1beta1/doc.go
@@ -16,7 +16,9 @@ limitations under the License.
 
 // +k8s:deepcopy-gen=package
 // +k8s:conversion-gen=k8s.io/kubernetes/pkg/kubelet/apis/config
-// +k8s:openapi-gen=true
+// +k8s:conversion-gen-external-types=k8s.io/kubelet/config/v1beta1
 // +k8s:defaulter-gen=TypeMeta
+// +k8s:defaulter-gen-input=../../../../../vendor/k8s.io/kubelet/config/v1beta1
+// +groupName=kubelet.config.k8s.io
 
-package v1beta1 // import "k8s.io/kubelet/config/v1beta1"
+package v1beta1 // import "k8s.io/kubernetes/pkg/kubelet/apis/config/v1beta1"
