+++ staging/src/k8s.io/kubelet/config/v1beta1/doc.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2017 The Kubernetes Authors.
+Copyright 2018 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -15,6 +15,7 @@ limitations under the License.
 */
 
 // +k8s:deepcopy-gen=package
-// TODO: This file should also enable openapi-gen like the other packages
+// +k8s:openapi-gen=true
+// +groupName=kubelet.config.k8s.io
 
-package config // import "k8s.io/kubernetes/pkg/kubelet/apis/config"
+package v1beta1 // import "k8s.io/kubelet/config/v1beta1"
